refactor: atualização na estrutura e nome de arquivos
Para manter a arquitetura proposta, foi feito uma adequação na estrutura das pastas. E foi ajustado o nome dos arquivos.
