+++ backend/api/routes/auth_routes.py
@@ -1,8 +1,10 @@
+# ID: backend/api/auth.py
 import random
 from fastapi import APIRouter, HTTPException, status, Depends
 from datetime import datetime, timedelta, timezone
 from fastapi.security import OAuth2PasswordBearer
-from schemas.auth import LoginRequest, LoginResponse, RegisterRequest, VerifyCodeRequest, ForgotPasswordRequest, ResetPasswordRequest
+from schemas.auth import LoginRequest, LoginResponse, VerifyCodeRequest, ForgotPasswordRequest, ResetPasswordRequest
+from schemas.user import RegisterRequest
 from services.auth_service import validar_data_nascimento
 from services.email_service import enviar_email_verificacao
 from database.config import supabase
@@ -175,8 +177,7 @@ async def reset_password(data: ResetPasswordRequest):
         "is_active": True
     }).eq("id", user["id"]).execute()
     return {"message": "Senha redefinida com sucesso! Você já pode fazer login."}
+
 @router.get("/usuarios/todos")
 async def listar_todos_usuarios(admin: dict = Depends(verificar_admin)):
-    # Se o código chegar aqui, é porque o usuário é admin.
-    # O objeto 'admin' contém os dados do token (id, email).
-    return {"status": "acesso autorizado"}
+    return {"status": "acesso authorized"}
